// src/types/jobs.ts
export interface Job {
  JobId: number;
  Ref: string;
  Type: string;
  Category: string;
  Contact: string;
  ContactId: number;
  ContactGoupId: number;
  Postcode: string;
  Location: string;
  JobContactLatitude: number;
  JobContactLongitude: number;
  Resource: string;
  Asset: string;
  PlannedStart: string;
  Duration: string;
  Status: string;
  StatusDate: string;
  RealStart: string;
  RealEnd: string;
  RealDuration: string;
  DrivingDur: string;
  Description: string;
  DrivingDist: number;
  PrintName: string;
  CustNote: string;
  ResNote: string;
  Actioned: string;
  JobPO: string | null;
  JobContractId: string | null;
  Created: string;
  Scheduled: string;
  CurrentFlag: string | null;
  Assistants: {
    ResourceId: number;
    ResourceName: string;
    Start?: string;
    End?: string;
  }[] | null;
  ContactParentId: number | null;
  StatusComment: string;
  JobTypeId: number;
  JobCategoryId: number;
}